Fix bug in StopsLayer where duplicate stations are not handled
Certain trains (e.g. those on the Ring Rail line) can have multiple time table row groups with the same station. This has to be handled properly when using the stations of these groups as a lookup map for maplibre expressions (there cannot be duplicates). In case of duplicate stations, choose the group based on the current time.