Show train composition and stopping position on each station on map
Adds a new layer TrainCompositionByStationLayer to the VehicleMapContainer. This new map layer component displays the train wagon composition and the exact stopping position of the train for each station where the train stops. The component utilizes new utility getTrainStoppingPositionOnTrack which utilizes generated static track sign data for each station platform in Finland. The generator utility for the track sign data is not yet present (to be added).